refactor(fbf): decouple home and quiz combo buttons
